+++ my/demo.py
@@ -23,7 +23,7 @@ class demo(user_config):
     def external_module(self):
         rpath = self.external
         if rpath is not None:
-            from .core.common import import_dir
+            from .core.utils.imports import import_dir
             return import_dir(rpath)
 
         import my.config.repos.external as m # type: ignore
